Revert "Add support for subcommand modules"
